:recycle: dms extract UseCaseType resolve to TargetResolverHelper
